requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
pyyaml>=5.4.0
tensorboard>=2.8.0
